BRMS-92-splunk-metrics: refactor code and update as per reivew comments.
