+++ app/uk/gov/hmrc/brm/models/brm/Payload.scala
@@ -35,22 +35,26 @@ case class Payload(
 
 object Payload extends BRMFormat {
 
+  val birthReferenceNumber = "birthReferenceNumber"
+  val firstName = "firstName"
+  val lastName = "lastName"
+  val dateOfBirth = "dateOfBirth"
   val whereBirthRegistered = "whereBirthRegistered"
 
   implicit val PayloadWrites: Writes[Payload] = (
-    (JsPath \ "birthReferenceNumber").write[Option[String]] and
-      (JsPath \ "firstName").write[String] and
-      (JsPath \ "lastName").write[String] and
-      (JsPath \ "dateOfBirth").write[LocalDate](jodaLocalDateWrites(datePattern)) and
+    (JsPath \ birthReferenceNumber).write[Option[String]] and
+      (JsPath \ firstName).write[String] and
+      (JsPath \ lastName).write[String] and
+      (JsPath \ dateOfBirth).write[LocalDate](jodaLocalDateWrites(datePattern)) and
       (JsPath \ whereBirthRegistered).write[BirthRegisterCountry](birthRegisterWrites)
 
     )(unlift(Payload.unapply))
 
   implicit val requestFormat: Reads[Payload] = (
-    (JsPath \ "birthReferenceNumber").readNullable[String](birthReferenceNumberValidate) and
-      (JsPath \ "firstName").read[String](minLength[String](1)) and
-      (JsPath \ "lastName").read[String](minLength[String](1)) and
-      (JsPath \ "dateOfBirth").read[LocalDate](isAfterDate) and
-      (JsPath \ "whereBirthRegistered").read[BirthRegisterCountry](birthRegisterReads)
+    (JsPath \ birthReferenceNumber).readNullable[String](birthReferenceNumberValidate) and
+      (JsPath \ firstName).read[String](minLength[String](1)) and
+      (JsPath \ lastName).read[String](minLength[String](1)) and
+      (JsPath \ dateOfBirth).read[LocalDate](isAfterDate) and
+      (JsPath \ whereBirthRegistered).read[BirthRegisterCountry](birthRegisterReads)
     )(Payload.apply _)
 }
